third_party/blink/web_tests/external/wpt/css/css-syntax/unicode-range-selector-expected.txt

This is a testharness.js-based test.
FAIL Unicode range is not a token, and doesn't cause confusion in selectors assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
Harness: the test ran to completion.
